+++ src/components/LogoutTimeout/LogoutTimeout.test.js
@@ -1,7 +1,10 @@
 import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import { userEvent } from '@folio/jest-config-stripes/testing-library/user-event';
+
 
 import LogoutTimeout from './LogoutTimeout';
 import { useStripes } from '../../StripesContext';
+import { getUnauthorizedPathFromSession, setUnauthorizedPathToSession } from '../../loginServices';
 
 
 jest.mock('../OrganizationLogo');
@@ -11,12 +14,28 @@ jest.mock('react-router', () => ({
 }));
 
 describe('LogoutTimeout', () => {
-  it('if not authenticated, renders a timeout message', async () => {
-    const mockUseStripes = useStripes;
-    mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false } });
+  describe('if not authenticated', () => {
+    it('renders a timeout message', async () => {
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false } });
 
-    render(<LogoutTimeout />);
-    screen.getByText('stripes-core.rtr.idleSession.sessionExpiredSoSad');
+      render(<LogoutTimeout />);
+      screen.getByText('stripes-core.rtr.idleSession.sessionExpiredSoSad');
+    });
+
+    it('clears previous path from storage after clicking', async () => {
+      const previousPath = '/monkey?bagel';
+      setUnauthorizedPathToSession(previousPath);
+      const user = userEvent.setup();
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false } });
+
+      render(<LogoutTimeout />);
+
+      await user.click(screen.getByRole('button'));
+
+      expect(getUnauthorizedPathFromSession()).toBe(null);
+    });
   });
 
   it('if authenticated, renders a redirect', async () => {